requirements.txt
flask
wheel
debugpy
jsonschema
attrs
beautifulsoup4
docopt
jinja2
jsonpointer
lxml
matplotlib
networkx
nltk
numpy
pandas
Pillow
pip
pydot
pyparsing
python-louvain
requests
setuptools
unidecode
xlrd
openpyxl
dogpile.cache
mpld3
python-epo-ops-client
sklearn
where
fuzzywuzzy
Flask
flask_cors
Sphinx
sphinx_rtd_theme
elasticsearch
hotreload

